py-data/requirements.txt
pyserial
matplotlib
websockets